# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Invoice & Receipt System API
  version: 0.1.0
  description: |
    Internal API for a small Israeli business operating as `עוסק פטור`.

    Conventions that hold everywhere:

    - Money is an **integer number of agorot** (1 ILS = 100 agorot). Never a
      decimal, never a float.
    - Timestamps are RFC 3339 in UTC. Business dates are `YYYY-MM-DD`, already
      interpreted in `Asia/Jerusalem`.
    - Authentication is a session cookie. Unsafe methods must also send the
      `X-CSRF-Token` header, echoing the `csrf_token` cookie.
    - Errors always use one envelope. `code` is a stable English identifier;
      `message` is Hebrew and is shown to the user as-is.

    Sending a document never affects the document: delivery runs in a
    background job, after issuance has committed, so a failed send cannot roll
    back an issued document (plan.md 14).

    Critical commands accept an `Idempotency-Key` header. A replay returns the
    original response with `Idempotent-Replay: true`, so a retry after a timeout
    never issues a second document with a second official number.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /expenses/:
    get:
      tags: [expenses]
      summary: List expenses, newest first
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - name: category
          in: query
          schema: { type: string }
        - name: activity_id
          in: query
          schema: { type: string, format: uuid }
        - name: payment_method
          in: query
          schema: { $ref: '#/components/schemas/PaymentMethod' }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of expenses
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExpensePage' }
    post:
      tags: [expenses]
      summary: Record an expense
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpenseInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
        '422': { $ref: '#/components/responses/Error' }
...
  /expenses/by-category:
    get:
      tags: [expenses]
      summary: Expenses summarized by category (plan.md 15)
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: activity_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Totals
          content:
            application/json:
              schema:
                type: object
                properties:
                  totals:
                    type: array
                    items: { $ref: '#/components/schemas/CategoryTotal' }
  /expenses/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [expenses]
      summary: One expense, with its attachments
      responses:
        '200':
          description: Expense
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      tags: [expenses]
      summary: Replace an expense
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpenseInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
        '422': { $ref: '#/components/responses/Error' }
    delete:
      tags: [expenses]
      summary: Delete an expense
      description: |
        Unlike an issued document, an expense is an internal record with no
        official number, so a mistake is corrected by deleting it. The deletion
        is audited, and the stored attachment files are retained.
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Error' }
  /expenses/{id}/attachments:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      tags: [expenses]
      summary: Attach a receipt photo or supplier PDF
      description: |
        The type is decided by sniffing the uploaded bytes, never by the
        filename or the declared content type, so a script named `receipt.pdf`
        is refused. Accepted: JPEG, PNG, HEIC, WEBP, PDF. Maximum 20 MiB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Stored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExpenseAttachment' }
        '413':
          description: The file is too large
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422':
          description: The bytes are not an accepted file type
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /expenses/{id}/attachments/{attachmentID}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - name: attachmentID
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [expenses]
      summary: Remove an attachment from an expense
      description: |
        Removes the record. The stored file itself is retained: storage is
        cheap, and an accidentally deleted receipt photo is not recoverable.
      responses:
        '204': { description: Removed }
  /expenses/attachments/{attachmentID}:
    parameters:
      - name: attachmentID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [expenses]
      summary: Download an attachment
      description: |
        Verified against its recorded SHA-256 and always served as an
        attachment with `nosniff`, so an uploaded file can never execute in
        this origin.
      responses:
        '200':
          description: The stored file
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/Error' }
...
  /exports/:
    post:
      tags: [exports]
      summary: Queue an accountant export
      description: |
        Builds one ZIP holding `documents.csv`, `payments.csv`, `expenses.csv`,
        `customers.csv`, a `manifest.json`, the issued PDFs under `pdf/<type>/`
        and the expense attachments (plan.md 15). CSVs are UTF-8 with a BOM so
        Excel opens Hebrew correctly.

        Building runs in the background; poll the status endpoint.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                from: { type: string, format: date }
                to: { type: string, format: date }
                include_pdfs: { type: boolean, default: true }
                include_attachments: { type: boolean, default: true }
      responses:
        '202':
          description: Queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string, format: uuid }
                  state: { type: string }
        '422': { $ref: '#/components/responses/Error' }
  /exports/{jobID}:
    parameters:
      - name: jobID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [exports]
      summary: How a build is going
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportStatus' }
        '404': { $ref: '#/components/responses/Error' }
  /exports/{jobID}/download:
    parameters:
      - name: jobID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [exports]
      summary: Download a finished archive
      description: Verified against the SHA-256 recorded when it was built.
      responses:
        '200':
          description: The archive
          content:
            application/zip:
              schema: { type: string, format: binary }
        '409':
          description: The build is not finished
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Set by `POST /auth/login`. Unsafe methods must also send
        `X-CSRF-Token`, echoing the `csrf_token` cookie.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    SearchQuery:
      name: q
      in: query
      schema: { type: string }
      description: Case-insensitive substring match
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable English machine identifier
              examples: [UNAUTHORIZED, VALIDATION_FAILED, DOCUMENT_IMMUTABLE]
            message:
              type: string
              description: Hebrew text, shown to the user as-is
            details:
              type: object
              additionalProperties: true
              description: Field-level messages, keyed by field name
            request_id:
              type: string
...
    PaymentMethod:
      type: string
      enum: [CASH, BANK_TRANSFER, CREDIT_CARD, BIT, PAYBOX, CHECK, OTHER]
...
    ExpenseAttachment:
      type: object
      properties:
        id: { type: string, format: uuid }
        original_filename:
          type: string
          description: What the uploader called it. It never determines where the file is written.
        content_type:
          type: string
          description: Determined by sniffing the bytes
        byte_size: { type: integer, format: int64 }
        sha256: { type: string }
        uploaded_at: { type: string, format: date-time }
    ExpenseInput:
      type: object
      required: [supplier, expense_date, amount_agorot]
      properties:
        supplier: { type: string, minLength: 1 }
        expense_date: { type: string, format: date }
        amount_agorot: { type: integer, format: int64, minimum: 1 }
        category: { type: string }
        payment_method:
          allOf: [{ $ref: '#/components/schemas/PaymentMethod' }]
          default: OTHER
        reference: { type: string }
        notes: { type: string }
        activity_id: { type: [string, 'null'], format: uuid }
    Expense:
      allOf:
        - $ref: '#/components/schemas/ExpenseInput'
        - type: object
          properties:
            id: { type: string, format: uuid, readOnly: true }
            payment_method_hebrew: { type: string, readOnly: true }
            activity_name: { type: [string, 'null'], readOnly: true }
            attachments:
              type: array
              readOnly: true
              items: { $ref: '#/components/schemas/ExpenseAttachment' }
            created_at: { type: string, format: date-time, readOnly: true }
            updated_at: { type: string, format: date-time, readOnly: true }
    ExpensePage:
      type: object
      properties:
        expenses:
          type: array
          items: { $ref: '#/components/schemas/Expense' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        total_amount_agorot: { type: integer, format: int64 }
    CategoryTotal:
      type: object
      properties:
        category: { type: string }
        count: { type: integer }
        total_agorot: { type: integer, format: int64 }
...
    ExportStatus:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        state:
          type: string
          enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        attempts: { type: integer }
        created_at: { type: string, format: date-time }
        filename: { type: string }
        bytes: { type: integer, format: int64 }
        sha256: { type: string }
        error: { type: string }
...